AI-CHATBOT-FORCOLLEGEWEBSIITE/Backend/requirements.txt
Flask
flask-cors
pymongo
python-dotenv
groq